Reflect doc links in bulk modify custom fields
